Test training application on success event
